move get_cbar_ticks from plotting/_slicers.py to plotting/_utils.py
